requirements.txt
gradio==5.49.1
numpy==1.21.5
paddleocr==2.7.0.0
paddleocr.egg==info
Pillow==12.0.0
torch==2.0.1
ultralytics==8.3.182
paddlepaddle==3.2.1
